+++ packages/cli/src/commands/versions/migrate.ts
@@ -23,7 +23,7 @@ import { resolve as resolvePath, join as joinPath } from 'path';
 import { OptionValues } from 'commander';
 import { readJson, writeJson } from 'fs-extra';
 import { minimatch } from 'minimatch';
-import { runInstall } from './bump';
+import { runInstall } from './install';
 import replace from 'replace-in-file';
 
 declare module 'replace-in-file' {
